Made necessary changes to integrate Night Manor into added world structure
