+++ worlds/ufo50/games/night_manor/__init__.py
@@ -1,0 +1,1 @@
+game_name = "Night Manor"
