+++ worlds/ufo50/games/night_manor/regions.py
@@ -1,21 +1,21 @@
-from typing import Dict, NamedTuple, TYPE_CHECKING
+from typing import Dict, NamedTuple, TYPE_CHECKING, List
 from BaseClasses import Region
 
-from .locations import create_night_manor_locations
-from .rules import create_night_manor_rules
+from .locations import create_locations
+from .rules import create_rules
 
 if TYPE_CHECKING:
-    from .. import UFO50World
+    from ... import UFO50World
 
 
 # not sure if we really need this yet, but making it in case we need it later since it's easy to remove
 class RegionInfo(NamedTuple):
-    pass
+    rooms: List[str] = []  # rooms this region contains, for the purpose of the garden prize access rule
 
 
 # keys are region names, values are the region object
 # for room names, the letter is the row (top to bottom), the number is the column (left to right)
-night_manor_region_info: Dict[str, RegionInfo] = {
+region_info: Dict[str, RegionInfo] = {
     "Starting Room": RegionInfo(), # the initial room the game starts in
     "First Floor & Exterior": RegionInfo(), # the floor accessible immediately after you exit the starting area
     "Second Floor": RegionInfo(), # second floor accessible after you get powered flashlight
@@ -25,15 +25,15 @@ night_manor_region_info: Dict[str, RegionInfo] = {
     "Basement": RegionInfo() #accessible after you get the iron key
 }
 
-
-def create_night_manor_regions_and_rules(world: "UFO50World") -> None:
+# this function is required, and its only argument can be the world class
+# it must return the regions that it created
+# it is recommended that you prepend each region name with the game it is from to avoid overlap
+def create_regions_and_rules(world: "UFO50World") -> Dict[str, Region]:
     night_manor_regions: Dict[str, Region] = {}
-    for region_name, region_data in night_manor_region_info.items():
-        night_manor_regions[region_name] = Region(
-            f"Night Manor - {region_name}", world.player, world.multiworld)
+    for region_name, region_data in region_info.items():
+        night_manor_regions[region_name] = Region(f"Night Manor - {region_name}", world.player, world.multiworld)
 
-    create_night_manor_locations(world, night_manor_regions)
-    create_night_manor_rules(world, night_manor_regions)
+    create_locations(world, night_manor_regions)
+    create_rules(world, night_manor_regions)
 
-    for region in night_manor_regions.values():
-        world.multiworld.regions.append(region)
+    return night_manor_regions
